+++ src/index.tsx
@@ -1,44 +1,8 @@
 import ReactDOM from "react-dom/client";
 import "@/index.css";
-import App from "@/App";
 import reportWebVitals from "./reportWebVitals";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import Home from "./pages/home/Home";
-import NotFound from "@/pages/NotFound";
-import Test from "./pages/test/test/Test";
-import TestResult from "@/pages/test/TestResult/TestResult";
-import BulletinBoard from "@/pages/board/BulletinBoard";
-import Stats from "@/pages/stats/Stats";
-import StatsMbti from "@/pages/stats/statsMbti/StatsMbti";
-import BoardPost from "@/components/board/BoardPost/BoardPost";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <App />,
-    errorElement: <NotFound />,
-    children: [
-      { index: true, path: "/", element: <Home /> },
-      { path: "/test", element: <Test /> },
-      { path: "/result", element: <TestResult /> },
-      { path: "/board", element: <BulletinBoard /> },
-      { path: "/board/:mbti", element: <BulletinBoard /> },
-      { path: "/stats", element: <Stats /> },
-      { path: "/stats/:mbti", element: <StatsMbti /> }
-    ]
-  },
-  {
-    // 게시글 작성 확인용 -> 추후에 BulletinBoard page 안에서 컴포넌트로만 사용할 예정
-    path: "/post",
-    element: (
-      <BoardPost
-        thisMbti={"ESTP"}
-        onThisClose={() => console.log("닫기")}
-        onThisComplete={(value) => console.log(value)}
-      />
-    )
-  }
-]);
+import { RouterProvider } from "react-router-dom";
+import router from "./routers";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
